Split up reducers.js.
